pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vpp"
version = "0.1.0"
description = "Virtual Power Plant RL environment for energy trading"
requires-python = ">=3.9"
dependencies = [
    "openenv-core",
    "pydantic>=2.0",
    "numpy>=1.24",
    "fastapi>=0.104",
    "uvicorn>=0.24",
    "httpx>=0.25",
    "websockets>=12.0",
    "python-dateutil>=2.8",
    "openai>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m vpp.server.app
server = "vpp.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["vpp", "vpp.server"]
package-dir = { "vpp" = ".", "vpp.server" = "server" }